Feb 9th, 2016 20hr40

+++ src/main/java/com/library/datamodel/jaxb/config/v1_0/Amounttype.java
@@ -43,7 +43,7 @@ import javax.xml.bind.annotation.XmlType;
 public class Amounttype {
 
     @XmlElement(required = true)
-    protected BigDecimal amount;
+    protected int amount;
     @XmlElement(required = true)
     protected String currencycode;
 
@@ -52,10 +52,10 @@ public class Amounttype {
      * 
      * @return
      *     possible object is
-     *     {@link BigDecimal }
+     *     {@link int }
      *     
      */
-    public BigDecimal getAmount() {
+    public int getAmount() {
         return amount;
     }
 
@@ -64,10 +64,10 @@ public class Amounttype {
      * 
      * @param value
      *     allowed object is
-     *     {@link BigDecimal }
+     *     {@link int }
      *     
      */
-    public void setAmount(BigDecimal value) {
+    public void setAmount(int value) {
         this.amount = value;
     }
